+++ CHANGES.rst
@@ -5,7 +5,7 @@ CHANGE HISTORY
 -----
 - Created the ParticleEffect and associated classes.
 - Implemented the StarFirework, RingFirework, SerpentFirework, PalmFirework,
-  Explosion and DropScreen effects.
+  Explosion, DropScreen and Rain effects.
 - Added background colour options to BarChart renderer.
 
 1.4.2